update  code highlighting
