--- src/components/SyntaxHighlighter.tsx
@@ -1,21 +1,0 @@
-import { PrismLight as SyntaxHighlighter } from 'react-syntax-highlighter'
-
-/**
- * hljs langs
- */
-import js from 'react-syntax-highlighter/dist/esm/languages/prism/javascript'
-import c from 'react-syntax-highlighter/dist/esm/languages/prism/c'
-import cpp from 'react-syntax-highlighter/dist/esm/languages/prism/cpp'
-import rust from 'react-syntax-highlighter/dist/esm/languages/prism/rust'
-import java from 'react-syntax-highlighter/dist/esm/languages/prism/java'
-import go from 'react-syntax-highlighter/dist/esm/languages/prism/go'
-
-SyntaxHighlighter.registerLanguage('java', java)
-SyntaxHighlighter.registerLanguage('javascript', js)
-SyntaxHighlighter.registerLanguage('c', c)
-SyntaxHighlighter.registerLanguage('cpp', cpp)
-SyntaxHighlighter.registerLanguage('cpp', cpp)
-SyntaxHighlighter.registerLanguage('rust', rust)
-SyntaxHighlighter.registerLanguage('go', go)
-
-export { SyntaxHighlighter }
